+++ src/controllers/console/client/Oauth2DeleteClientAction.php
@@ -4,13 +4,14 @@ namespace rhertogh\Yii2Oauth2Server\controllers\console\client;
 
 use rhertogh\Yii2Oauth2Server\controllers\console\client\base\Oauth2BaseClientAction;
 use rhertogh\Yii2Oauth2Server\controllers\console\Oauth2ClientController;
+use rhertogh\Yii2Oauth2Server\interfaces\controllers\console\client\Oauth2DeleteClientActionInterface;
 use yii\console\ExitCode;
 use yii\helpers\Console;
 
 /**
  * @property Oauth2ClientController $controller
  */
-class Oauth2DeleteClientAction extends Oauth2BaseClientAction
+class Oauth2DeleteClientAction extends Oauth2BaseClientAction implements Oauth2DeleteClientActionInterface
 {
     public function run($id)
     {