+++ src/interfaces/controllers/console/encryption/Oauth2RotateEncryptionKeysActionInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace rhertogh\Yii2Oauth2Server\interfaces\controllers\console\encryption;
+
+interface Oauth2RotateEncryptionKeysActionInterface
+{
+}